Don't initialize anything before setup() is called.
I learned this lesson the hard way on another project (and then forgot
it): initializing a C++ object (like Graphics gfx; at the top of
main.cpp) before the system can initialize causes a race. If the wrong
task wins the race, malloc() doesn't work and the whole system crashes.